5. move classes to main package
