--- src/main/java/com/together/buytogether/domain/Address.java
@@ -1,15 +1,0 @@
-package com.together.buytogether.domain;
-
-import jakarta.persistence.*;
-
-@Entity
-public class Address {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "address_id")
-    Long id;
-
-    @Column(unique = true)
-    String postCode;
-}
