+++ src/main/java/com/together/buytogether/member/domain/Member.java
@@ -1,4 +1,4 @@
-package com.together.buytogether.member.feature;
+package com.together.buytogether.member.domain;
 
 import org.springframework.util.Assert;
