--- src/main/java/com/together/buytogether/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package com.together.buytogether.repository;
-
-import com.together.buytogether.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByLoginId(String loginId);
-}
